add istoria subcomponent
